🛠 피버 타임 카운트다운 리팩토링
- 카운트다운 로직 DanggnGameController로 이동
